% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decentralized.R
\name{print.DecentralizedFedGEE}
\alias{print.DecentralizedFedGEE}
\title{Print a decentralized FedGEE fit}
\usage{
\method{print}{DecentralizedFedGEE}(x, ...)
}
\arguments{
\item{x}{A \code{DecentralizedFedGEE} fit.}

\item{...}{Unused.}
}
\description{
Print a decentralized FedGEE fit
}
