% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedgee.R
\name{print.FedGEE}
\alias{print.FedGEE}
\title{Print Method for Federated GEE}
\usage{
\method{print}{FedGEE}(x, ...)
}
\arguments{
\item{x}{An object of class \code{FedGEE}.}

\item{...}{Passed to \code{\link{summary.FedGEE}}.}
}
\description{
Print Method for Federated GEE
}
