% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedgee.R
\name{vcov.FedGEE}
\alias{vcov.FedGEE}
\title{Variance-covariance matrix of a FedGEE fit}
\usage{
\method{vcov}{FedGEE}(object, correction = object$correction, ...)
}
\arguments{
\item{object}{A \code{FedGEE} fit.}

\item{correction}{Which stored variant; defaults to the fit's choice.}

\item{...}{Unused.}
}
\description{
Variance-covariance matrix of a FedGEE fit
}
